requirements.txt
supabase==1.0.3
python-telegram-bot==20.5
yfinance==0.2.31
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pykrx==1.0.40
schedule==1.2.0
python-dotenv==1.0.0
pytz==2023.3
tqdm==4.65.0 